# openapi/AuthService.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AlgoPath Auth Service",
    "version": "1.0.0",
    "description": "Authentication, accounts, subscriptions, billing, notifications, and social APIs.\n\nDirect service endpoint on its internal port. Public gateway prefix: /api/auth."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Internal service instance"
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/webhooks/cashfree": {
      "post": {
        "tags": [
          "Billing"
        ],
        "summary": "Create or execute action on webhooks cashfree",
        "description": "Handles POST requests for /api/v1/auth/webhooks/cashfree. Implemented in server/AuthService/src/server.ts:77.",
        "operationId": "createAuthWebhooksCashfree",
        "security": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operation succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "x-source": {
          "file": "server/AuthService/src/server.ts",
          "line": 77
        },
        "requestBody": {
          "description": "JSON request payload processed by controller at server/AuthService/src/server.ts:77.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/webhooks/stripe": {
      "post": {
        "tags": [
          "Billing"
        ],
        "summary": "Create or execute action on webhooks stripe",
        "description": "Handles POST requests for /api/v1/auth/webhooks/stripe. Implemented in server/AuthService/src/server.ts:72.",
        "operationId": "createAuthWebhooksStripe",
        "security": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operation succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "x-source": {
          "file": "server/AuthService/src/server.ts",
          "line": 72
        },
        "requestBody": {
          "description": "JSON request payload processed by controller at server/AuthService/src/server.ts:72.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token issued by AuthService."
      },
      "internalServiceSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-secret",
        "description": "Configured service-to-service secret."
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "description": "Standard response wrapper returned across AlgoPath REST endpoints.",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          },
          "data": {
            "description": "Payload data object or array."
          },
          "meta": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "success",
          "message"
        ],
        "additionalProperties": true
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response payload.",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ],
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Validation or operation failure"
          },
          "code": {
            "type": "string",
            "example": "INVALID_INPUT"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        },
        "required": [
          "success",
          "message"
        ]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request body, query parameter, or validation error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "Request state conflicts with existing resource.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Unexpected internal server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}